correccion en idiomas

+++ vistas/multi_form_action.php
@@ -92,13 +92,13 @@ if (isset($_SESSION['id_user']) && isset($_SESSION['id_rol'])) {
 			$habilidades = validarString($_POST["habilidades"]);
 			//$lastlogin = "2021-11-21 01:30:11";
 
-			if (isset($_FILES["pdf"])) {
+			/*if (isset($_FILES["pdf"])) {
 				$pdf = $_FILES['pdf']['name'];
 				$temp = $_FILES['pdf']['tmp_name'];
 				move_uploaded_file($temp, 'cv/' . $pdf);
 			} else {
 				$pdf = NULL;
-			}
+			}*/
 
 			//$licencia = validarNum($_POST["licencia"]);
 
@@ -131,16 +131,20 @@ if (isset($_SESSION['id_user']) && isset($_SESSION['id_rol'])) {
 
 				//comprueba si no hay idiomas guardados
 				if (mysqli_num_rows($datos) == 0) {
+					echo "<br>si el us no tiene idiomas guardados entra aca<br>";
 					//se guardan los idiomas ingresados
 					for ($y = 0; $y < count($idiomas); $y++) {
 						$aux = $idiomas[$y];
 						$x = mysqli_query($conexion, "INSERT INTO idioxuser (iduser, idi) VALUES ('$idloc','$aux')");
 					}
 				} else {
+					echo "<br>si tiene idiomas guardados<br>";
 					$b = 0;
-					// For de los idiomas ingresados  
+					// For de los idiomas ingresados  aca esta el problem
+					echo "Cantidad de idiomas ingresados: " . count($idiomas)  . "<br>";
 					for ($y = 0; $y < count($idiomas); $y++) {
 						while ($fila = mysqli_fetch_row($datos)) {
+							echo "Idioma ingresado: " . $idiomas[$y] . " " . "Idioma en bd: " . $fila[0] . "<br>";
 							if ($idiomas[$y] == $fila[0]) {
 								$b = 1;
 							}
@@ -150,8 +154,11 @@ if (isset($_SESSION['id_user']) && isset($_SESSION['id_rol'])) {
 							$x = mysqli_query($conexion, "INSERT INTO idioxuser (iduser, idi) VALUES ('$idloc','$aux')");
 						} else {
 							$b = 0;
+							$datos = mysqli_query($conexion, "SELECT idi FROM idioxuser WHERE iduser='$idloc'");
 						}
 					}
+
+
 					$datos = mysqli_query($conexion, "SELECT idi FROM idioxuser WHERE iduser='$idloc'");
 
 					while ($fila = mysqli_fetch_row($datos)) {
@@ -199,50 +206,82 @@ if (isset($_SESSION['id_user']) && isset($_SESSION['id_rol'])) {
 				$niveledu = "";
 			}
 
-			if (isset($_FILES["foto"])) {
-
-
+			if (isset($_FILES["foto"]) && $_FILES["foto"]['name'] != '') {
+				echo "entra en ISSET";
 				$fotobd = mysqli_query($conexion, "SELECT foto FROM usuario WHERE idloc='$idloc'");
 				$row = mysqli_fetch_array($fotobd);
-				if (mysqli_num_rows($fotobd) != 0) {
-					if(unlink('images/'.$row[0])) {
+				echo "aca" . $row[0] . "<----<br>";
+
+				if ($row[0]) {
+					if (unlink('images/' . $row[0])) {
 						echo "bien";
 						// file was successfully deleted
-					  } else {
+					} else {
+						echo "mal";
 						// there was a problem deleting the file
-					  }
+					}
 				}
+
 				$foto = $_FILES['foto']['name'];
+				echo "foto<br>" . $foto;
 				$temp = $_FILES['foto']['tmp_name'];
 				if (move_uploaded_file($temp, "images/" . $foto)) {
 					//Cambiamos los permisos del archivo a 777 para poder modificarlo posteriormente
-					chmod('images/' . $foto, 0777);
+					//chmod('images/' . $foto, 0777);
 				} else {
+					echo "<br>no se guardo en carpeta<br>";
+				}
+			} else {
+				echo "<br>cuando no cargas foto entra aca<br>";
+				$fotobd = mysqli_query($conexion, "SELECT foto FROM usuario WHERE idloc='$idloc'");
+				$row = mysqli_fetch_array($fotobd);
+				if($row[0]){
+					$foto = $row[0];
+				}else{
 					$foto = NULL;
 				}
+				
 			}
 
 
 
-			if (isset($_FILES["pdf"])) {
-				$pdf = $_FILES['pdf']['name'];
-				//Si el archivo contiene algo y es diferente de vacio
+			if (isset($_FILES["pdf"]) && $_FILES["pdf"]['name'] != '') {
+				
+				$cvbd = mysqli_query($conexion, "SELECT pdf FROM usuario WHERE idloc='$idloc'");
+				$rowcv = mysqli_fetch_array($cvbd);
 
+				if ($rowcv[0]) {
+					if (unlink('cv/' . $rowcv[0])) {
+						// file was successfully deleted
+					} else {
+						// there was a problem deleting the file
+					}
+				}
+				//Si el archivo contiene algo y es diferente de vacio
 				//Obtenemos algunos datos necesarios sobre el archivo
 				//$tipo = $_FILES['pdf']['type'];
 				//$tamano = $_FILES['pdf']['size'];
+				$pdf = $_FILES['pdf']['name'];
 				$temp = $_FILES['pdf']['tmp_name'];
 				if (move_uploaded_file($temp, 'cv/' . $pdf)) {
 					//Cambiamos los permisos del archivo a 777 para poder modificarlo posteriormente
-					chmod('images/' . $pdf, 0777);
+					//chmod('images/' . $pdf, 0777);
 					//Mostramos el mensaje de que se ha subido co éxito
-
+				} else {
+					echo "no se guardo en carpeta<br>";
 				}
 			} else {
-				$pdf = NULL;
+				$cvbd = mysqli_query($conexion, "SELECT pdf FROM usuario WHERE idloc='$idloc'");
+				$rowcv = mysqli_fetch_array($cvbd);
+
+				if ($rowcv[0]) {
+					$pdf = $rowcv[0];
+				}else{
+					$pdf = NULL;
+				}
+				
 			}
 
-			// Variable id del usuario
 
 			$y = mysqli_query($conexion, "SELECT * FROM usuario WHERE idloc='$idloc'");
 			if (mysqli_num_rows($y) == 0) {
